+++ rullst-capital/src/lib.rs
@@ -6,7 +6,7 @@ pub mod fiscal;
 pub mod invoice;
 pub mod providers;
 
-#[cfg(feature = "axum")]
+#[cfg(any(feature = "axum", feature = "actix"))]
 pub mod webhook;
 
 pub use billable::*;
@@ -16,5 +16,5 @@ pub use error::*;
 pub use fiscal::*;
 pub use invoice::*;
 
-#[cfg(feature = "axum")]
+#[cfg(any(feature = "axum", feature = "actix"))]
 pub use webhook::*;